+++ packages/gatsby-plugin-postcss/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@2.1.21...gatsby-plugin-postcss@2.1.22) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-postcss
+
 ## [2.1.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@2.1.20...gatsby-plugin-postcss@2.1.21) (2020-03-06)
 
 **Note:** Version bump only for package gatsby-plugin-postcss