+++ packages/gatsby-theme-blog-core/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.27...gatsby-theme-blog-core@1.2.28) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.26...gatsby-theme-blog-core@1.2.27) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.25...gatsby-theme-blog-core@1.2.26) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.24...gatsby-theme-blog-core@1.2.25) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.23...gatsby-theme-blog-core@1.2.24) (2020-03-12)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.22...gatsby-theme-blog-core@1.2.23) (2020-03-12)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.21...gatsby-theme-blog-core@1.2.22) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.20...gatsby-theme-blog-core@1.2.21) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.19...gatsby-theme-blog-core@1.2.20) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.2.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.18...gatsby-theme-blog-core@1.2.19) (2020-03-10)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
 ## [1.2.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.2.17...gatsby-theme-blog-core@1.2.18) (2020-03-10)
 
 **Note:** Version bump only for package gatsby-theme-blog-core